@@ -1,17 +1,18 @@
 {
   "name": "link-dir",
   "version": "1.0.3",
-  "description": "Cross-platform directory linking",
-  "main": "dist/index.js",
+  "description": "Link directory",
+  "main": "lib/index.js",
   "files": [
-    "dist/"
+    "lib"
   ],
-  "typings": "dist/index.d.ts",
-  "bin": "dist/cli.js",
+  "typings": "lib/index.d.ts",
+  "bin": "lib/cli.js",
   "keywords": [
     "link",
     "linking",
-    "junction",
+    "hardlink",
+    "hardlinking",
     "directory",
     "dir",
     "folder"
@@ -40,14 +41,16 @@
     "cash-rm": "^0.2.0",
     "mos": "^1.3.1",
     "mos-plugin-readme": "^1.0.4",
-    "typescript": "^2.0.3"
+    "typescript": "^2.2.0"
   },
   "dependencies": {
     "@types/mz": "0.0.30",
     "@types/node": "^6.0.45",
-    "mkdirp": "^0.5.1",
-    "mkdirp-promise": "^3.0.1",
-    "mz": "^2.4.0"
+    "bole": "^3.0.2",
+    "graceful-fs": "^4.1.11",
+    "mkdirp-promise": "^5.0.1",
+    "mz": "^2.4.0",
+    "rimraf-then": "^1.0.1"
   },
   "mos": {
     "plugins": [
