@@ -1,25 +1,53 @@
-import symlink, {SymlinkType} from './forceSymlink'
+import {Stats} from 'fs';
 import path = require('path')
-import os = require('os')
-import mkdirp = require('mkdirp-promise/lib/node4')
+import fs = require('mz/fs')
+import mkdirp = require('mkdirp-promise')
+import rimraf = require('rimraf-then')
+import bole = require('bole')
 
-// Always use "junctions" on Windows. Even though support for "symbolic links" was added in Vista+, users by default
-// lack permission to create them
-const symlinkType: SymlinkType = os.platform() === 'win32' ? 'junction' : 'dir'
+const logger = bole('link-dir')
 
-/**
- * Relative symlink
- */
-export default async function relSymlink (src: string, dest: string) {
-  // Junction points can't be relative
-  const rel = symlinkType !== 'junction' ? path.relative(path.dirname(dest), src) : src
+export default async function (existingDir: string, newDir: string) {
+  const stage = `${newDir}+stage`
+  await rimraf(stage)
+  await hardlinkDir(existingDir, stage)
+  await rimraf(newDir)
+  await fs.rename(stage, newDir)
+}
 
+async function hardlinkDir(existingDir: string, newDir: string) {
+  await mkdirp(newDir)
+  const dirs = await fs.readdir(existingDir)
+  await Promise.all(
+    dirs
+      .map(async (relativePath: string) => {
+        const existingPath = path.join(existingDir, relativePath)
+        const newPath = path.join(newDir, relativePath)
+        const stat: any = await fs.lstat(existingPath)
+        if (stat.isSymbolicLink() || stat.isFile()) {
+          return safeLink(existingPath, newPath, stat)
+        }
+        if (stat.isDirectory()) {
+          return hardlinkDir(existingPath, newPath)
+        }
+      })
+  )
+}
+
+async function safeLink(existingPath: string, newPath: string, stat: Stats) {
   try {
-    await symlink(rel, dest, symlinkType)
+    await fs.link(existingPath, newPath)
   } catch (err) {
-    if ((<NodeJS.ErrnoException>err).code === 'ENOENT') {
-      await mkdirp(path.dirname(dest))
-      await symlink(rel, dest, symlinkType)
+    // shouldn't normally happen, but if the file was already somehow linked,
+    // the installation should not fail
+    if (err.code === 'EEXIST') {
+      return
+    }
+    // might happen if package contains a broken symlink, we don't fail on this
+    if (err.code === 'ENOENT' && stat.isSymbolicLink()) {
+      logger.warn({
+        message: `Broken symlink found: ${existingPath}`,
+      })
       return
     }
     throw err
