@@ -1,4 +1,14 @@
-declare module 'mkdirp-promise/lib/node4' {
+declare module 'bole' {
+  const anything: any;
+  export = anything;
+}
+
+declare module 'mkdirp-promise' {
+  const anything: any;
+  export = anything;
+}
+
+declare module 'rimraf-then' {
   const anything: any;
   export = anything;
 }
